Refactor form handling: remove FormProvider, add Auto, Endgame, and Teleop components, and update state management
