+++ src/routes/scouting/$matchId/collect/pre-match.tsx
@@ -1,17 +1,6 @@
-import { SegmentedControl } from "@/components/inputs";
-import { useMatch } from "@/data/db";
-import {
-  Alliance,
-  DrivePosition,
-  MatchType,
-  preMatchDefaults,
-  preMatchSchema,
-} from "@/data/match/shared";
+import { PreMatch } from "@/components/form/PreMatch";
 import { useAppState } from "@/data/state";
-import { useForm } from "@/hooks/useForm";
-import { enumToSelectItem } from "@/util";
 import { createFileRoute } from "@tanstack/react-router";
-import { useEffect, useState } from "react";
 
 export const Route = createFileRoute("/scouting/$matchId/collect/pre-match")({
   component: Page,
@@ -21,56 +10,6 @@ export const Route = createFileRoute("/scouting/$matchId/collect/pre-match")({
 });
 
 function Page(): JSX.Element {
-  const [allianceColor, setAllianceColor] = useState<
-    "red" | "blue" | undefined
-  >();
   const { matchId } = Route.useParams();
-  const match = useMatch(matchId);
-
-  useEffect(() => {
-    setAllianceColor(
-      match?.phases.preMatch?.alliance || preMatchDefaults.alliance,
-    );
-  }, [match]);
-
-  const form = useForm<"preMatch">({
-    matchId,
-    phase: "preMatch",
-    initialValues: match?.phases.preMatch || preMatchDefaults,
-    schema: preMatchSchema,
-  });
-
-  form.watch("alliance", ({ value }) => {
-    setAllianceColor(value);
-  });
-
-  return (
-    <div className="grid md:grid-cols-2 gap-x-4 gap-y-2">
-      <SegmentedControl
-        className="col-span-full"
-        label="Alliance"
-        data={enumToSelectItem(Alliance)}
-        fullWidth
-        color={allianceColor}
-        {...form.getInputProps("alliance")}
-      />
-
-      <SegmentedControl
-        className="col-span-full"
-        label="Match Type"
-        data={enumToSelectItem(MatchType)}
-        fullWidth
-        color={allianceColor}
-        {...form.getInputProps("matchType")}
-      />
-
-      <SegmentedControl
-        className="col-span-full"
-        label="Drive Position"
-        data={enumToSelectItem(DrivePosition)}
-        color={allianceColor}
-        {...form.getInputProps("drivePosition")}
-      />
-    </div>
-  );
+  return <PreMatch matchId={matchId} />;
 }
